add authentication in auth module
