+++ Backend/src/auth/auth.module.ts
@@ -1,0 +1,61 @@
+import { Module } from '@nestjs/common';
+import { AuthController } from './auth.controller';
+import { AuthService } from './services/auth.service';
+import { EjsAdapter } from '@nestjs-modules/mailer/dist/adapters/ejs.adapter';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from 'src/user/entity/user.entity';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import jwtConfig from './config/jwt.config';
+import { JwtModule } from '@nestjs/jwt';
+import { MailerModule } from '@nestjs-modules/mailer';
+import { join } from 'path';
+import { HashingProvider } from './services/hashing.provider';
+import { GenerateTokenProvider } from 'src/common/services/generate-token.provider';
+import { UploadService } from 'src/common/services/upload.service';
+import { MailService } from './services/mail.service';
+import { UserRepository } from 'src/user/repositories/user.repository';
+
+@Module({
+  imports: [
+    TypeOrmModule.forFeature([User]),
+    ConfigModule.forFeature(jwtConfig),
+    JwtModule.registerAsync(jwtConfig.asProvider()),
+    MailerModule.forRootAsync({
+      inject: [ConfigService],
+      useFactory: async (config: ConfigService) => ({
+        transport: {
+          host: config.get('appConfig.mailHost'),
+          secure: false,
+          port: 2525,
+          auth: {
+            user: config.get('appConfig.smtpUsername'),
+            pass: config.get('appConfig.smtpPassword'),
+          },
+        },
+        default: {
+          from: `twitter <no-reply@twitter.com>`,
+        },
+        template: {
+          dir: join(__dirname, 'templates'),
+          adapter: new EjsAdapter({
+            inlineCssEnabled: true,
+          }),
+          options: {
+            strict: false,
+          },
+        },
+      }),
+    }),
+  ],
+
+  controllers: [AuthController],
+  providers: [
+    UserRepository,
+    AuthService,
+    HashingProvider,
+    GenerateTokenProvider,
+    UploadService,
+    MailService,
+  ],
+})
+export class AuthModule {}
