+++ Backend/src/auth/dtos/google-token.dto.ts
@@ -1,0 +1,6 @@
+import { IsNotEmpty } from 'class-validator';
+
+export class GoogleTokenDto {
+  @IsNotEmpty()
+  token: string;
+}
