+++ Backend/src/auth/templates/verify-email.ejs
