+++ Backend/src/user/user.module.ts
@@ -3,57 +3,25 @@ import { User } from './entity/user.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserController } from './user.controller';
 import { UserService } from './services/user.service';
-import { HashingProvider } from './services/hashing.provider';
-import { ConfigModule, ConfigService } from '@nestjs/config';
+import { ConfigModule } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
-import { GenerateTokenProvider } from './services/generate-token.provider';
 import { UploadService } from '../common/services/upload.service';
-import { MailService } from './services/mail.service';
-import jwtConfig from './config/jwt.config';
-import { MailerModule } from '@nestjs-modules/mailer';
-import { join } from 'path';
-import { EjsAdapter } from '@nestjs-modules/mailer/dist/adapters/ejs.adapter';
+import jwtConfig from '../auth/config/jwt.config';
+
 import { UserRepository } from './repositories/user.repository';
+import { GenerateTokenProvider } from 'src/common/services/generate-token.provider';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([User]),
     ConfigModule.forFeature(jwtConfig),
     JwtModule.registerAsync(jwtConfig.asProvider()),
-    MailerModule.forRootAsync({
-      inject: [ConfigService],
-      useFactory: async (config: ConfigService) => ({
-        transport: {
-          host: config.get('appConfig.mailHost'),
-          secure: false,
-          port: 2525,
-          auth: {
-            user: config.get('appConfig.smtpUsername'),
-            pass: config.get('appConfig.smtpPassword'),
-          },
-        },
-        default: {
-          from: `twitter <no-reply@twitter.com>`,
-        },
-        template: {
-          dir: join(__dirname, 'templates'),
-          adapter: new EjsAdapter({
-            inlineCssEnabled: true,
-          }),
-          options: {
-            strict: false,
-          },
-        },
-      }),
-    }),
   ],
   controllers: [UserController],
   providers: [
     UserService,
-    HashingProvider,
     GenerateTokenProvider,
     UploadService,
-    MailService,
     UserRepository,
   ],
   exports: [UserModule, UserRepository],
